Update README.md
Overview of Challenge Project - Note the rubric is clear as to where points are granted or not but it isn't clear as to what constitutes a good write-up. An example would have been helpful. Also knowing in advance what tasks are to be recorded (notes in readme file and screenshots of work) would have been extremely helpful and made for a more fair grading process. You can't go back and screenshot your process after the fact. I also wonder what screenshots would be helpful that don't contain things included in the material. That isn't a sarcastic criticism. I am really interested in what is expected in this regard.